# cookie parse: add fast path