Important change is at line 70, if we don't have any `=` left in the string, no need to continue the loop as we do not have any key-value pairs left